bugfix, wrong rate was read

+++ src/main/java/com/github/morotsman/beverage/product_catalog/service/ProductCatalogServiceImpl.java
@@ -48,7 +48,7 @@ public class ProductCatalogServiceImpl implements ProductCatalogService {
 
     @Override
     public Optional<Product> getProduct(Long id) {
-        return productRepository.findOne(id.byteValue());
+        return productRepository.findOne(id.longValue());
     }
